Make entry ordering in undo script deterministic
Using the POSIX locale for directory entry sorting guarantees the an
ordering by ASCII codepoints independent of the current locale.

Also make the command which deletes the script file the last command of
the undo script, because otherwise it is not clear whether the remaining
commands will still be executed or not.